Earlier use of fix precision math eariler in rake calc to avoid rounding errors

+++ server/poker-game/src/main/java/com/cubeia/games/poker/adapter/FirebaseServerAdapter.java
@@ -518,7 +518,7 @@ public class FirebaseServerAdapter implements ServerAdapter {
 	@Override
 	public void notifyRakeInfo(RakeInfoContainer rakeInfoContainer) {
 	    log.debug("--> sending rake info to client: {}", rakeInfoContainer);
-		RakeInfo rakeInfo = new RakeInfo(rakeInfoContainer.getTotalPot(), rakeInfoContainer.getTotalRake());
+		RakeInfo rakeInfo = new RakeInfo((int) rakeInfoContainer.getTotalPot(), (int) rakeInfoContainer.getTotalRake());
 		GameDataAction action = protocolFactory.createGameAction(rakeInfo, 0, table.getId());
 		sendPublicPacket(action, -1);
 	}
